Some refactor on the variable structure and debug-tag
The jail_ prefix is gone.
Now jail_foo should be defined under the jail config item as a dict key.
Like this:
jail:
  foo: bar

This structural change goes also for the host and the child jails as well.
Also a custom option appeared in the jail side. Whatever you define here,
that goes straight to the jail.conf as a value.

Debug messages only appear on explicit request:
User --tags all,debug

Also, mount support improved: Now you don't need the boilerplate.
You can define a list of directory. A list item can be a pair of directories
in case the directory name differs on the host and within the guest.