freeform inet configuration for jail interfaces
This feature is for the cases, when you don't want / can not follow the
eXb_hostname nameing convention.
In these cases you can use freeform inet configuration.
If you wish, you can mix the the. NB.: Using the freeform inet configuration
won't step the counter for the classic "way". You can arrange those dicts to
any part of the list.